3 Content Scripts working
I am not entirely sure what was going on. I used three content scripts in order to do the same thing as 'background.js' in the main Verallel extension since background scripts are deprecated on Firefox. It works, but anything that needs to be added to background.js will likely have to be made into either a content script or using another method like tabs.executeScript()